requirements.txt
click>=8.0.0
GitPython>=3.1.0
openai>=1.0.0
anthropic>=0.18.0
pyyaml>=6.0
rich>=13.0.0
jinja2>=3.0.0